+++ sn_node_manager/src/lib.rs
@@ -774,6 +774,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -876,6 +878,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -943,6 +947,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -1053,6 +1059,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -1133,6 +1141,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -1223,6 +1233,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -1312,6 +1324,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -1371,6 +1385,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -1420,6 +1436,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -1467,6 +1485,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -1517,6 +1537,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -1580,6 +1602,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -1706,6 +1730,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -1794,6 +1820,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -1927,6 +1955,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -2072,6 +2102,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -2212,6 +2244,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -2353,6 +2387,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -2524,6 +2560,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -2678,6 +2716,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: Some(LogFormat::Json),
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -2835,6 +2875,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -2989,6 +3031,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             number: 1,
             node_ip: Some(Ipv4Addr::new(192, 168, 1, 1)),
@@ -3146,6 +3190,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             number: 1,
             node_ip: None,
@@ -3186,6 +3232,321 @@ mod tests {
             .await?;
 
         assert_eq!(service_manager.service.service_data.node_port, Some(12000));
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn upgrade_should_retain_max_archived_log_files() -> Result<()> {
+        let current_version = "0.1.0";
+        let target_version = "0.2.0";
+
+        let tmp_data_dir = assert_fs::TempDir::new()?;
+        let current_install_dir = tmp_data_dir.child("safenode_install");
+        current_install_dir.create_dir_all()?;
+
+        let current_node_bin = current_install_dir.child("safenode");
+        current_node_bin.write_binary(b"fake safenode binary")?;
+        let target_node_bin = tmp_data_dir.child("safenode");
+        target_node_bin.write_binary(b"fake safenode binary")?;
+
+        let mut mock_service_control = MockServiceControl::new();
+        let mut mock_rpc_client = MockRpcClient::new();
+
+        // before binary upgrade
+        mock_service_control
+            .expect_get_process_pid()
+            .with(eq(current_node_bin.to_path_buf().clone()))
+            .times(1)
+            .returning(|_| Ok(1000));
+        mock_service_control
+            .expect_stop()
+            .with(eq("safenode1"), eq(false))
+            .times(1)
+            .returning(|_, _| Ok(()));
+
+        // after binary upgrade
+        mock_service_control
+            .expect_uninstall()
+            .with(eq("safenode1"), eq(false))
+            .times(1)
+            .returning(|_, _| Ok(()));
+        mock_service_control
+            .expect_install()
+            .with(
+                eq(ServiceInstallCtx {
+                    args: vec![
+                        OsString::from("--rpc"),
+                        OsString::from("127.0.0.1:8081"),
+                        OsString::from("--root-dir"),
+                        OsString::from("/var/safenode-manager/services/safenode1"),
+                        OsString::from("--log-output-dest"),
+                        OsString::from("/var/log/safenode/safenode1"),
+                        OsString::from("--max-archived-log-files"),
+                        OsString::from("20"),
+                    ],
+                    autostart: false,
+                    contents: None,
+                    environment: None,
+                    label: "safenode1".parse()?,
+                    program: current_node_bin.to_path_buf(),
+                    username: Some("safe".to_string()),
+                    working_directory: None,
+                }),
+                eq(false),
+            )
+            .times(1)
+            .returning(|_, _| Ok(()));
+
+        // after service restart
+        mock_service_control
+            .expect_start()
+            .with(eq("safenode1"), eq(false))
+            .times(1)
+            .returning(|_, _| Ok(()));
+        mock_service_control
+            .expect_wait()
+            .with(eq(3000))
+            .times(1)
+            .returning(|_| ());
+        mock_service_control
+            .expect_get_process_pid()
+            .with(eq(current_node_bin.to_path_buf().clone()))
+            .times(1)
+            .returning(|_| Ok(100));
+
+        mock_rpc_client.expect_node_info().times(1).returning(|| {
+            Ok(NodeInfo {
+                pid: 2000,
+                peer_id: PeerId::from_str("12D3KooWS2tpXGGTmg2AHFiDh57yPQnat49YHnyqoggzXZWpqkCR")?,
+                data_path: PathBuf::from("/var/safenode-manager/services/safenode1"),
+                log_path: PathBuf::from("/var/log/safenode/safenode1"),
+                version: target_version.to_string(),
+                uptime: std::time::Duration::from_secs(1), // the service was just started
+                wallet_balance: 0,
+            })
+        });
+        mock_rpc_client
+            .expect_network_info()
+            .times(1)
+            .returning(|| {
+                Ok(NetworkInfo {
+                    connected_peers: Vec::new(),
+                    listeners: Vec::new(),
+                })
+            });
+
+        let mut service_data = NodeServiceData {
+            auto_restart: false,
+            connected_peers: None,
+            data_dir_path: PathBuf::from("/var/safenode-manager/services/safenode1"),
+            genesis: false,
+            home_network: false,
+            listen_addr: None,
+            local: false,
+            log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
+            log_format: None,
+            max_archived_log_files: Some(20),
+            max_log_files: None,
+            metrics_port: None,
+            node_ip: None,
+            node_port: None,
+            number: 1,
+            owner: None,
+            peer_id: Some(PeerId::from_str(
+                "12D3KooWS2tpXGGTmg2AHFiDh57yPQnat49YHnyqoggzXZWpqkCR",
+            )?),
+            pid: Some(1000),
+            reward_balance: Some(AttoTokens::zero()),
+            rpc_socket_addr: SocketAddr::new(IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)), 8081),
+            safenode_path: current_node_bin.to_path_buf(),
+            service_name: "safenode1".to_string(),
+            status: ServiceStatus::Running,
+            upnp: false,
+            user: Some("safe".to_string()),
+            user_mode: false,
+            version: current_version.to_string(),
+        };
+        let service = NodeService::new(&mut service_data, Box::new(mock_rpc_client));
+
+        let mut service_manager = ServiceManager::new(
+            service,
+            Box::new(mock_service_control),
+            VerbosityLevel::Normal,
+        );
+
+        service_manager
+            .upgrade(UpgradeOptions {
+                auto_restart: false,
+                bootstrap_peers: Vec::new(),
+                env_variables: None,
+                force: false,
+                start_service: true,
+                target_bin_path: target_node_bin.to_path_buf(),
+                target_version: Version::parse(target_version).unwrap(),
+            })
+            .await?;
+
+        assert_matches!(
+            service_manager.service.service_data.max_archived_log_files,
+            Some(20)
+        );
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn upgrade_should_retain_max_log_files() -> Result<()> {
+        let current_version = "0.1.0";
+        let target_version = "0.2.0";
+
+        let tmp_data_dir = assert_fs::TempDir::new()?;
+        let current_install_dir = tmp_data_dir.child("safenode_install");
+        current_install_dir.create_dir_all()?;
+
+        let current_node_bin = current_install_dir.child("safenode");
+        current_node_bin.write_binary(b"fake safenode binary")?;
+        let target_node_bin = tmp_data_dir.child("safenode");
+        target_node_bin.write_binary(b"fake safenode binary")?;
+
+        let mut mock_service_control = MockServiceControl::new();
+        let mut mock_rpc_client = MockRpcClient::new();
+
+        // before binary upgrade
+        mock_service_control
+            .expect_get_process_pid()
+            .with(eq(current_node_bin.to_path_buf().clone()))
+            .times(1)
+            .returning(|_| Ok(1000));
+        mock_service_control
+            .expect_stop()
+            .with(eq("safenode1"), eq(false))
+            .times(1)
+            .returning(|_, _| Ok(()));
+
+        // after binary upgrade
+        mock_service_control
+            .expect_uninstall()
+            .with(eq("safenode1"), eq(false))
+            .times(1)
+            .returning(|_, _| Ok(()));
+        mock_service_control
+            .expect_install()
+            .with(
+                eq(ServiceInstallCtx {
+                    args: vec![
+                        OsString::from("--rpc"),
+                        OsString::from("127.0.0.1:8081"),
+                        OsString::from("--root-dir"),
+                        OsString::from("/var/safenode-manager/services/safenode1"),
+                        OsString::from("--log-output-dest"),
+                        OsString::from("/var/log/safenode/safenode1"),
+                        OsString::from("--max-log-files"),
+                        OsString::from("20"),
+                    ],
+                    autostart: false,
+                    contents: None,
+                    environment: None,
+                    label: "safenode1".parse()?,
+                    program: current_node_bin.to_path_buf(),
+                    username: Some("safe".to_string()),
+                    working_directory: None,
+                }),
+                eq(false),
+            )
+            .times(1)
+            .returning(|_, _| Ok(()));
+
+        // after service restart
+        mock_service_control
+            .expect_start()
+            .with(eq("safenode1"), eq(false))
+            .times(1)
+            .returning(|_, _| Ok(()));
+        mock_service_control
+            .expect_wait()
+            .with(eq(3000))
+            .times(1)
+            .returning(|_| ());
+        mock_service_control
+            .expect_get_process_pid()
+            .with(eq(current_node_bin.to_path_buf().clone()))
+            .times(1)
+            .returning(|_| Ok(100));
+
+        mock_rpc_client.expect_node_info().times(1).returning(|| {
+            Ok(NodeInfo {
+                pid: 2000,
+                peer_id: PeerId::from_str("12D3KooWS2tpXGGTmg2AHFiDh57yPQnat49YHnyqoggzXZWpqkCR")?,
+                data_path: PathBuf::from("/var/safenode-manager/services/safenode1"),
+                log_path: PathBuf::from("/var/log/safenode/safenode1"),
+                version: target_version.to_string(),
+                uptime: std::time::Duration::from_secs(1), // the service was just started
+                wallet_balance: 0,
+            })
+        });
+        mock_rpc_client
+            .expect_network_info()
+            .times(1)
+            .returning(|| {
+                Ok(NetworkInfo {
+                    connected_peers: Vec::new(),
+                    listeners: Vec::new(),
+                })
+            });
+
+        let mut service_data = NodeServiceData {
+            auto_restart: false,
+            connected_peers: None,
+            data_dir_path: PathBuf::from("/var/safenode-manager/services/safenode1"),
+            genesis: false,
+            home_network: false,
+            listen_addr: None,
+            local: false,
+            log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
+            log_format: None,
+            max_archived_log_files: None,
+            max_log_files: Some(20),
+            metrics_port: None,
+            node_ip: None,
+            node_port: None,
+            number: 1,
+            owner: None,
+            peer_id: Some(PeerId::from_str(
+                "12D3KooWS2tpXGGTmg2AHFiDh57yPQnat49YHnyqoggzXZWpqkCR",
+            )?),
+            pid: Some(1000),
+            reward_balance: Some(AttoTokens::zero()),
+            rpc_socket_addr: SocketAddr::new(IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)), 8081),
+            safenode_path: current_node_bin.to_path_buf(),
+            service_name: "safenode1".to_string(),
+            status: ServiceStatus::Running,
+            upnp: false,
+            user: Some("safe".to_string()),
+            user_mode: false,
+            version: current_version.to_string(),
+        };
+        let service = NodeService::new(&mut service_data, Box::new(mock_rpc_client));
+
+        let mut service_manager = ServiceManager::new(
+            service,
+            Box::new(mock_service_control),
+            VerbosityLevel::Normal,
+        );
+
+        service_manager
+            .upgrade(UpgradeOptions {
+                auto_restart: false,
+                bootstrap_peers: Vec::new(),
+                env_variables: None,
+                force: false,
+                start_service: true,
+                target_bin_path: target_node_bin.to_path_buf(),
+                target_version: Version::parse(target_version).unwrap(),
+            })
+            .await?;
+
+        assert_matches!(service_manager.service.service_data.max_log_files, Some(20));
 
         Ok(())
     }
@@ -3300,6 +3661,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: Some(12000),
             node_ip: None,
             node_port: None,
@@ -3457,6 +3820,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: Some(12000),
             node_ip: None,
             node_port: None,
@@ -3614,6 +3979,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -3771,6 +4138,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -3927,6 +4296,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -3999,6 +4370,8 @@ mod tests {
             local: false,
             log_dir_path: log_dir.to_path_buf(),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -4056,6 +4429,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -4129,6 +4504,8 @@ mod tests {
             local: false,
             log_dir_path: PathBuf::from("/var/log/safenode/safenode1"),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -4194,6 +4571,8 @@ mod tests {
             local: false,
             log_dir_path: log_dir.to_path_buf(),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,
@@ -4257,6 +4636,8 @@ mod tests {
             local: false,
             log_dir_path: log_dir.to_path_buf(),
             log_format: None,
+            max_archived_log_files: None,
+            max_log_files: None,
             metrics_port: None,
             node_ip: None,
             node_port: None,